requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.26.109
botocore==1.29.109
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-import-export==2.8.0
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
jmespath==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
ratelim==0.1.6
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
toml==0.10.2
tzdata==2023.3
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
